+++ src/main/java/plugily/projects/villagedefense/handlers/language/LanguageMigrator.java
@@ -1,6 +1,6 @@
 /*
  * Village Defense - Protect villagers from hordes of zombies
- * Copyright (C) 2020  Plugily Projects - maintained by 2Wild4You, Tigerpanzer_02 and contributors
+ * Copyright (C) 2021  Plugily Projects - maintained by 2Wild4You, Tigerpanzer_02 and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -37,8 +37,8 @@ import java.util.List;
 @SuppressWarnings("deprecation")
 public class LanguageMigrator {
 
-  public static final int LANGUAGE_FILE_VERSION = 16;
-  public static final int CONFIG_FILE_VERSION = 14;
+  public static final int LANGUAGE_FILE_VERSION = 17;
+  public static final int CONFIG_FILE_VERSION = 15;
   private final Main plugin;
   private final List<String> migratable = Arrays.asList(Constants.Files.CONFIG.getName(), Constants.Files.KITS.getName(),
           Constants.Files.KITS.getName(), Constants.Files.LANGUAGE.getName(), Constants.Files.SPECIAL_ITEMS.getName(), Constants.Files.MYSQL.getName());
@@ -177,6 +177,12 @@ public class LanguageMigrator {
                   "# Default: true\r\n" +
                   "InGame-Join-Respawn: true\r\n" +
                   "\r\n");
+          break;
+        case 14:
+          MigratorUtils.addNewLines(file, "\r\n" +
+                  "# Can the players buy again iron golems or wolves if these\r\n" +
+                  "# entities died? The config limit and permission will be ignored." +
+                  "Players-Can-Buy-GolemsWolves-If-They-Died: false\r\n\r\n");
           break;
         default:
           break;
@@ -367,6 +373,11 @@ public class LanguageMigrator {
           MigratorUtils.insertAfterLine(file, "Admin-Commands:", "    Spychat-Command:");
           MigratorUtils.insertAfterLine(file, "Spychat-Command:", "      Toggled: \"&aGame spy chat toggled to&c %value%\"");
           break;
+        case 15:
+          //??? MISSED ???
+          break;
+        case 16:
+          MigratorUtils.insertAfterLine(file, "  Item:", "    Name: \"&f%mapname%\"");
         default:
           break;
       }